+++ commands/reset_log.go
@@ -3,6 +3,7 @@ package commands
 import (
 	"fmt"
 
+	multiplexer "git.randomchars.net/Reviath/RemiliaScarlet/Multiplexer"
 	CommandHandler "git.randomchars.net/Reviath/RemiliaScarlet/handler"
 	"git.randomchars.net/Reviath/RemiliaScarlet/sql"
 )
@@ -17,12 +18,15 @@ func ResetLogCommand(ctx CommandHandler.Context, _ []string) error {
 	var tag Tag
 
 	if sql.CheckLanguage(ctx.Guild.ID) == "tr" {
+		if !multiplexer.CheckAdministratorPermission(ctx.Session, ctx.Message.Author.ID, ctx.Channel.ID) {
+			ctx.Reply("Yeterli yetkiye sahip değilsin.")
+			return nil
+		}
 		err := db.QueryRow(fmt.Sprintf("SELECT channelid FROM log WHERE guildid ='%s'", ctx.Guild.ID)).Scan(&tag.channelid)
 		if err == nil {
 			delete, err := db.Query(fmt.Sprintf("DELETE FROM log WHERE guildid ='%s'", ctx.Guild.ID))
 			if err != nil {
 				ctx.Reply("Bir hata oluştu.")
-
 				return nil
 			}
 
@@ -37,6 +41,12 @@ func ResetLogCommand(ctx CommandHandler.Context, _ []string) error {
 			return nil
 		}
 	}
+
+	if !multiplexer.CheckAdministratorPermission(ctx.Session, ctx.Message.Author.ID, ctx.Channel.ID) {
+		ctx.Reply("You don't have enough permission.")
+		return nil
+	}
+
 	err := db.QueryRow(fmt.Sprintf("SELECT channelid FROM log WHERE guildid ='%s'", ctx.Guild.ID)).Scan(&tag.channelid)
 	if err == nil {
 		delete, err := db.Query(fmt.Sprintf("DELETE FROM log WHERE guildid ='%s'", ctx.Guild.ID))
